chore(PERM-3): leaf-extract F1/F3/F4 helpers into sibling modules to satisfy File Size Budget
Pure refactor, zero behavior change: the F1/F3/F4 additions pushed three files a
few lines over their architecture File Size Budget. Extract the added helpers into
new same-directory sibling modules (imported back), no budget raised.

- ipc-parsing.ts (771->under): stripShellCommandEnvPrefix + SHELL_TOOL_NAMES ->
  ipc-shell-command-prefix.ts.
- auto-permission-read-only-gate.ts (712->under): known-safe catalog + git/sed/find
  guards -> auto-permission-read-only-catalog.ts.
- tool-execution-policy-service.ts (711->651): resolveCapabilityRules + type +
  once-log set -> tool-execution-capability-resolution.ts.

check:architecture File Size Budget passes; format:check + typecheck clean; 211
affected tests pass. Extraction done via Codex (companion task --write).
