+++ apps/core/src/runtime/ipc-parsing.ts
@@ -15,13 +15,13 @@ import {
 import { parseIpcMessageFiles } from './ipc-message-files.js';
 import { parseSemanticCapabilityDefinitionsRecord } from '../shared/semantic-capabilities.js';
 import { isPlainObject, toTrimmedString } from '../shared/object.js';
-import { stripHostInjectedEnvPrefix } from '../shared/runtime-env-command.js';
 import {
   validateBrowserIpcAuthRequest,
   validateIpcAuthRequest,
   validateMemoryIpcAuthRequest,
 } from './ipc-auth-validation.js';
 import { parseInteractionDescriptor } from './ipc-interaction-descriptor-parsing.js';
+import { stripShellCommandEnvPrefix } from './ipc-shell-command-prefix.js';
 import { sanitizeIpcToolInput } from './ipc-tool-input-sanitization.js';
 import { PERMISSION_CLASSIFIER_MAX_STRING_LENGTH } from './permission-classifier-prompt.js';
 
@@ -305,33 +305,6 @@ export function parseMemoryIpcRequest(
       : {}),
   };
 }
-const SHELL_TOOL_NAMES = new Set(['Bash', 'RunCommand']);
-
-/**
- * For shell tools, return a shallow-copied toolInput with the host-injected env
- * prefix stripped from the `command`/`cmd` field. Non-shell tools and inputs
- * without a string command field pass through unchanged.
- */
-function stripShellCommandEnvPrefix(
-  toolName: string,
-  toolInput: unknown,
-): unknown {
-  if (!SHELL_TOOL_NAMES.has(toolName) || !isPlainObject(toolInput)) {
-    return toolInput;
-  }
-  const field =
-    typeof toolInput.command === 'string'
-      ? 'command'
-      : typeof toolInput.cmd === 'string'
-        ? 'cmd'
-        : undefined;
-  if (!field) return toolInput;
-  return {
-    ...toolInput,
-    [field]: stripHostInjectedEnvPrefix(toolInput[field] as string).command,
-  };
-}
-
 export function parsePermissionIpcRequest(
   raw: unknown,
   sourceAgentFolder: string,
