+++ apps/core/src/shared/auto-permission-read-only-gate.ts
@@ -7,6 +7,18 @@ import {
   type BashCommandLeaf,
   type BashCommandParseResult,
 } from './bash-command-parser.js';
+import {
+  BARE_SAFE_EXECUTABLES,
+  capabilityTokens,
+  FIND_GLOBAL_OPTION,
+  FIND_UNSAFE_PRIMARY,
+  GENERIC_READ_EXECUTABLES,
+  genericReadFileArgs,
+  gitReadOnly,
+  hasHiddenPathSegment,
+  normalizeCapabilityId,
+  sedReadFileArgs,
+} from './auto-permission-read-only-catalog.js';
 import { mcpToolPatternCovers } from './mcp-tool-scope.js';
 import { allProtectedPathMentions } from './tool-execution-protected-paths.js';
 
@@ -60,54 +72,6 @@ const LS_LONG_OPTIONS =
 // individually. Redirects (`<`/`>`), command substitution (`` ` ``/`$(...)`),
 // braces, globs (`*`/`?`/`[]`), and comments (`#`) still block outright.
 const SHELL_CONTROL_OR_EXPANSION = /[\r\n#<>`$(){}*?\[\]]/;
-// Argument-shape-agnostic reads that never touch a workspace file (they print
-// literals, identity, or transform stdin). Any file/secret they might name is
-// still caught by the whole-command protected-path and per-leaf secret scans.
-// `env`/`printenv` are deliberately EXCLUDED: they dump the process
-// environment, which can hold live credentials — auto-allowing them would
-// contradict the gate that blocks reading `.env`.
-const BARE_SAFE_EXECUTABLES = new Set([
-  'basename',
-  'date',
-  'dirname',
-  'echo',
-  'expr',
-  'false',
-  'id',
-  'seq',
-  'tr',
-  'true',
-  'uname',
-  'whoami',
-]);
-// Read stdin or a file operand; confined the same way as cat/grep.
-const GENERIC_READ_EXECUTABLES = new Set([
-  'cut',
-  'nl',
-  'paste',
-  'rev',
-  'sort',
-  'uniq',
-]);
-const GIT_READ_ONLY_SUBCOMMANDS = new Set([
-  'branch',
-  'diff',
-  'log',
-  'show',
-  'status',
-]);
-const GIT_SAFE_GLOBAL_OPTIONS = new Set([
-  '--literal-pathspecs',
-  '--no-optional-locks',
-  '--no-pager',
-  '--no-replace-objects',
-]);
-const GIT_BRANCH_READ_OPTIONS =
-  /^(?:-a|-r|-v|-vv|--list|--all|--remotes|--verbose|--color(?:=\w+)?|--no-color)$/;
-// -exec/-delete/etc. run or mutate; -follow/-L/-H escape workspace confinement.
-const FIND_UNSAFE_PRIMARY =
-  /^-(?:exec|execdir|okdir|ok|delete|fprintf|fprint0|fprint|fls|follow|L|H)$/;
-const FIND_GLOBAL_OPTION = /^-(?:O\d*|P|D|f)$/;
 const SECRET_KEY =
   /(?:^|[_-])(?:apikey|authorization|credential|key|password|private[_-]?key|secret|token)(?:$|[_-])/i;
 const SECRET_PATH =
@@ -292,68 +256,6 @@ function evaluateLeaf(
   return blocked(
     `Executable ${executable || '(missing)'} is not a reviewed read command.`,
   );
-}
-
-// Read-only git subset (status/log/diff/show/read-only branch). Rejects `-c`,
-// `-C`, `--git-dir`, `--exec-path`, output/exec options, every write/network
-// subcommand, and any hidden/secret pathspec (e.g. `git show HEAD:.npmrc`).
-function gitReadOnly(
-  args: readonly string[],
-  capabilityIds: readonly string[],
-): boolean {
-  if (!capabilityIds.some((id) => capabilityTokens(id)[0] === 'git')) {
-    return false;
-  }
-  let index = 0;
-  while (index < args.length && args[index]!.startsWith('-')) {
-    if (!GIT_SAFE_GLOBAL_OPTIONS.has(args[index]!)) return false;
-    index += 1;
-  }
-  const subcommand = args[index];
-  if (!subcommand || !GIT_READ_ONLY_SUBCOMMANDS.has(subcommand)) return false;
-  const rest = args.slice(index + 1);
-  for (const arg of rest) {
-    if (arg === '-c' || /^(?:-o|-O|--output|--exec)/.test(arg)) return false;
-    const pathPart = arg.includes(':') ? arg.slice(arg.indexOf(':') + 1) : arg;
-    if (hasHiddenPathSegment(pathPart)) return false;
-  }
-  if (subcommand === 'branch') {
-    return rest.every((arg) => GIT_BRANCH_READ_OPTIONS.test(arg));
-  }
-  return true;
-}
-
-// Only `sed -n <script> [file...]` (print-only). Any other flag (`-i`, `-e`,
-// `-f`, `--in-place`) blocks, as does a script naming a write/exec command.
-function sedReadFileArgs(args: readonly string[]): string[] | undefined {
-  const flags = args.filter((arg) => arg.startsWith('-'));
-  if (!flags.includes('-n') || flags.some((flag) => flag !== '-n')) {
-    return undefined;
-  }
-  const nonFlags = args.filter((arg) => !arg.startsWith('-'));
-  const script = nonFlags[0] ?? '';
-  // Conservative: a script char of w/W (write) or e (execute) blocks, even when
-  // it is only regex text — the safe cost is an extra prompt.
-  if (/[wWe]/.test(script)) return undefined;
-  return nonFlags.slice(1);
-}
-
-// stdin-or-file transforms: skip option flags, reject output flags, confine the
-// rest. Over-rejects separated option VALUES (they resolve to non-files and
-// block), which only costs an extra prompt.
-function genericReadFileArgs(args: readonly string[]): string[] | undefined {
-  const fileArgs: string[] = [];
-  let optionsEnded = false;
-  for (const arg of args) {
-    if (!optionsEnded && arg === '--') {
-      optionsEnded = true;
-    } else if (!optionsEnded && arg.startsWith('-')) {
-      if (/^(?:-o|-w|--output)/.test(arg)) return undefined;
-    } else {
-      fileArgs.push(arg);
-    }
-  }
-  return fileArgs;
 }
 
 // parseBashCommand rejects `find` outright, so vet a single (non-compound)
@@ -632,13 +534,6 @@ function blockedReadShape(
   return blocked(`The file ${action} command shape is not provably safe.`);
 }
 
-function hasHiddenPathSegment(value: string): boolean {
-  return value
-    .replaceAll('\\', '/')
-    .split('/')
-    .some((segment) => segment !== '.' && segment.startsWith('.'));
-}
-
 function isWithinPath(base: string, candidate: string): boolean {
   const relative = path.relative(base, candidate);
   return (
@@ -693,14 +588,6 @@ function commandText(input: unknown): string | undefined {
   const record = input as Record<string, unknown>;
   const value = record.command ?? record.cmd;
   return typeof value === 'string' && value.trim() ? value.trim() : undefined;
-}
-
-function normalizeCapabilityId(value: string): string {
-  return value.trim().toLowerCase().replaceAll(/[_-]+/g, '.');
-}
-
-function capabilityTokens(value: string): string[] {
-  return normalizeCapabilityId(value).split('.').filter(Boolean);
 }
 
 function allowed(reason: string): AutoPermissionReadOnlyGateResult {
